Fetch uptime from Twitch streams API
- (Test to ensure that `user_id: user` works properly)
- Move uptime logic from internal-stats to twitch
- Remove unnecessary argument passed through to internal-stats endpoint
