+++ index.js
@@ -27,7 +27,7 @@ app.get('/callback-spotify', initSpotify.callback)
 app.get('/callback-twitch', initTwitch.callback)
 app.get('/spotify.json', require('./endpoints/json-spotify'))
 app.get('/twitch.json', require('./endpoints/json-twitch'))
-app.get('/internal-stats.json', require('./endpoints/json-internal-stats')( { startTime }))
+app.get('/internal-stats.json', require('./endpoints/json-internal-stats'))
 app.get('/splatoon.json', require('./endpoints/json-splatoon'))
 app.get('/acnh.json', require('./endpoints/json-acnh'))
 app.get('/chat', require('./endpoints/chat')({ startTime }))
